% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{distinct_color_ramp}
\alias{distinct_color_ramp}
\title{Generate distint colors}
\usage{
distinct_color_ramp(n)
}
\arguments{
\item{n}{The number of colors to generate}
}
\value{
a character vector of colors
}
\description{
Generate distint colors
}
\examples{
distinct_color_ramp(4)
}
