% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_enrichment.R
\name{enrichment_gene_list}
\alias{enrichment_gene_list}
\title{Annotates list of features using g:OSt}
\usage{
enrichment_gene_list(
  list_features,
  organism,
  n.terms = 3,
  user_threshold = 0.01,
  recall = 0.2,
  precision = 0.2,
  ...
)
}
\arguments{
\item{list_features}{list of list containing groups of related features/genes}

\item{organism}{Organism name, constructed by concatenating the fist letter of the
name with the family name}

\item{n.terms}{Number of terms to save}

\item{user_threshold}{p-value threshold for significance}

\item{recall}{percent of intersect that overlap with reference terms}

\item{precision}{percent of intersect that overlap with query terms}

\item{...}{Additional arguments passed to \link[gprofiler2]{gost}}
}
\value{
A list with the significant term names and associated p-values
respectively, ordered by significance.
}
\description{
Enriches list of features/genes with most significant terms according
to g:OSt by using \link[gprofiler2]{gost}.
}
\examples{
#Markers for proliferation and EMT transition
genes <- list(c("MKI67", "TOP2A", "MYBL2", "BUB1"), c("CA9", "VEGFA", "ZEB1", "VIM", "FN1"))
enrichment_gene_list(genes, organism = "hsapiens")

}
