% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{gg_color}
\alias{gg_color}
\title{Generate ggplot2 colors}
\usage{
gg_color(n)
}
\arguments{
\item{n}{The number of colors to generate}
}
\value{
a character vector of colors
}
\description{
reference John Colby's stack overflow response
}
\examples{
gg_color(4)
}
